refactor: rename relation_manager folder
