+++ moqt-server/src/lib.rs
@@ -3,7 +3,7 @@ use crate::modules::{
     buffer_manager,
     buffer_manager::{buffer_manager, BufferCommand},
     control_message_handler::*,
-    relation_manager::{commands::PubSubRelationCommand, manager::pubsub_relation_manager},
+    pubsub_relation_manager::{commands::PubSubRelationCommand, manager::pubsub_relation_manager},
     send_stream_dispatcher::{send_stream_dispatcher, SendStreamDispatchCommand},
 };
 use anyhow::{bail, Context, Ok, Result};
@@ -282,7 +282,7 @@ async fn handle_connection(
 
     // Delete pub/sub information related to the client
     let pubsub_relation_manager =
-        modules::relation_manager::interface::PubSubRelationManagerInterface::new(
+        modules::pubsub_relation_manager::interface::PubSubRelationManagerInterface::new(
             track_namespace_tx.clone(),
         );
     let _ = pubsub_relation_manager.delete_client(stable_id).await;
@@ -329,7 +329,7 @@ async fn handle_incoming_bi_stream(
     let shread_send_stream = &mut stream.shread_send_stream;
 
     let mut pubsub_relation_manager =
-        modules::relation_manager::interface::PubSubRelationManagerInterface::new(
+        modules::pubsub_relation_manager::interface::PubSubRelationManagerInterface::new(
             track_namespace_tx.clone(),
         );
     let mut send_stream_dispatcher =
